Refresh token when running cron
Before accessing a repo during a cron job the users token is
refreshed like it is done for e.g. create pipeline. User in this context means the creator of the cron job.

If the scheduled delay between two executions of a cron job is larger than the exiration time of the oauth2 refresh token (provided by the forge) this will fail and the cron job will not be run. But this is the best one can do.