Forge Github Org: Use `FullPath` instead of `Name`
`Name` is the human readable name, which could be "Woodpecker CI"
 for the `Path` "woodpecker-ci".
Gitlab supports subgroups, so we need to use the FullPath, otherwise we
would only get `sub-group` when owner is `parent-group/sub-group`.

Fixes #2092 when using Gitlab.

This was causing the organization lookup to fail, because it looks up using the `login`, when it did not find the organization, it would try to create it. The creation would fail, because it uses the `name`, and an organization with that `name` already exists.
Resulting in
```
pq: duplicate key value violates unique constraint "UQE_orgs_name"
```