docs: missing info for runs_on (#2457)
The `runs_on` section is missing from the syntax docs. So I added a
section for it, and linked to an example.

It's good enough. But I think it might be a good idea for someone (with
more knowledge than me) to add more info in this section. How to use it,
when to use it, etc. The example is useful but not very informative.