RegisterAgent with hostname
Transfer the agent hostname to the server's model.agent.Name field.

Previously the autoscaler had to create the agent with the server and
get a unique token to pass to the agent at boot up.
This allows to get the right name for the agent while using the shared
master agent token at agent boot up.
Both ways are still supported.