Fix error container overflow (#2957)
Fixes: https://github.com/woodpecker-ci/woodpecker/issues/2947


![image](https://github.com/woodpecker-ci/woodpecker/assets/3391958/03198aec-fd3c-4fcd-8418-a7c5b0ff9d0b)

On the mobile view, it now wraps to show error on top. That is still not
perfect as it creates content jumps, after clicking on a pipeline with
errors. I don't have a better idea yet, but IMO it's already an
improvement as before it was quite unusable on mobile view.

Before:

![image](https://github.com/woodpecker-ci/woodpecker/assets/3391958/20849de8-55d6-4839-b4b4-fe220003887d)

After:

![image](https://github.com/woodpecker-ci/woodpecker/assets/3391958/8a80939b-d6a5-414d-b693-ef4583e2f37d)