[Awesome Woodpecker] added yet another autoscaler (#3011)
Hello there :wave: 

we tried a few different autoscaler implementations along out in k8s
runners on arm, but most of them do not play well with a mix of build
agents. so we just build our own.
maybe some other folks find it beneficial so i would propose to list it
here as well.