Read datasource config from file when run in docker
The location of the sqlite3 database file is defined in a config file.
The location of the config file is provided to the woodpecker-server in
an ENV var.
This is done instead of providing the location directly inside an ENV var, as
the config file has the least precedence, so it can be overriden in a
docker-compose.yml file.

Remember: The config file must not have a line break at the end!