Documentation: kubernetes
Removed `settings` from the volume example, as it isn't needed, and throws an error in 2.7.3

Added a link to the Kubernetes PVC documentation, and a note about ReadWriteMany.

Added details around setting up prometheus monitoring where both prometheus and woodpecker are in the same cluster.

Included troubleshooting for prometheus failing to scrape metrics.