Add branch filter to cli pipeline purge
This patch adds a `--branch` option to `woodpecker-cli pipeline purge` to enable purging only the pipeline for one branch, while keeping the other pipelines intact.

Use case would be a the pipelines of a merged and possibly deleted development branch, while keeping the pipelines of the release branch intact.

Also, make the `older-than` parameter optional, since having to specify `--older-than 1s` is not useful.

Signed-off-by: Oliver Eikemeier <eikemeier@fillmore-labs.com>