fix(bitbucketdatacenter): fix CI_COMMIT_PULL_REQUEST (#5768)
The CI_COMMIT_PULL_REQUEST environment variable was not filled out for
Bitbucket Datacenter edition. This was because this ID comes from the
`ref`, which for this forge was set to the branch name of the
pull-request, and not the reference that contains the pull-request ID.

The pullRequest.fromRef.ID is "refs/heads/branch-name".

Bitbucket datacenter also publishes this as "refs/pull-requests/123/from"

This change switches from using the ref from the event payload to the
one that contains the pull-request ID in the path, allowing the
CI_COMMIT_PULL_REQUEST variable to be set correctly.